requirements.txt
# Solana
solders>=0.21.0
base58>=2.1.1

# HTTP
aiohttp>=3.9.0

# Env
python-dotenv>=1.0.0
